requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
boto3==1.26.53
botocore==1.29.53
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-storages==1.13.2
idna==3.4
jmespath==1.0.1
mysqlclient==2.1.1
numpy==1.24.1
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.14
xmltodict==0.13.0
